[nrf noup] Add Zephyr module manifest in the root directory
Add Zephyr module manifest in the CHIP root directory in
order to automatically pull CHIP into a project when CHIP
is one of the submodules specified in west manifest.

CHIP module takes all compilatinon flags associated with
the zephyr_interface target and passes them to the external
GN-based project. For that reason we must ensure that the
Matter module is processed after other revelant modules
that modify the zephyr_interface target. nrfxlib is such
a module as it contains nrf_security which carries the
mbedtls configuration.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>